pyproject.toml
[project]
name = "kseal"
version = "1.0.1"
description = "A kubeseal companion CLI - decrypt, export, and encrypt Kubernetes Secrets with automatic binary management"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Bruno Bernard", email = "contact.brunobernard@gmail.com" },
]
license = { text = "MIT" }
keywords = ["kubernetes", "kubeseal", "sealed-secrets", "secrets", "cli"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "kubernetes>=28.0",
    "ruamel-yaml>=0.18",
    "httpx>=0.25",
    "rich>=14.2.0",
    "packaging>=23.0",
    "pydantic>=2.0",
]

[project.scripts]
kseal = "kseal.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kseal"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.9",
    "basedpyright>=1.29.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["kseal"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]

[tool.basedpyright]
typeCheckingMode = "standard"
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportAny = false
reportExplicitAny = false
reportMissingTypeStubs = false
